+++ src/debug_utils.h
@@ -1,0 +1,33 @@
+/* Copyright (C) 2021 Greenbone Networks GmbH
+ *
+ * SPDX-License-Identifier: AGPL-3.0-or-later
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+
+/**
+ * @file debug_utils.h
+ * @brief Headers for debug utilties and Sentry integration
+ */
+
+#ifndef _OPENVAS_DEBUG_UTILS_H
+#define _OPENVAS_DEBUG_UTILS_H
+
+#include <gvm/base/gvm_sentry.h> /* for gvm_sentry_init */
+
+int
+init_sentry (void);
+
+#endif