+++ src/manage.c
@@ -46,6 +46,7 @@
  */
 #define _GNU_SOURCE
 
+#include "debug_utils.h"
 #include "gmp_base.h"
 #include "manage.h"
 #include "manage_acl.h"
@@ -79,6 +80,7 @@
 #include <time.h>
 #include <unistd.h>
 
+#include <gvm/base/gvm_sentry.h>
 #include <gvm/base/hosts.h>
 #include <gvm/base/proctitle.h>
 #include <gvm/osp/osp.h>
@@ -2478,7 +2480,7 @@ launch_osp_openvas_task (task_t task, target_t target, const char *scan_id,
 {
   osp_connection_t *connection;
   char *hosts_str, *ports_str, *exclude_hosts_str, *finished_hosts_str;
-  gchar *clean_hosts, *clean_exclude_hosts;
+  gchar *clean_hosts, *clean_exclude_hosts, *clean_finished_hosts_str;
   int alive_test, reverse_lookup_only, reverse_lookup_unify;
   osp_target_t *osp_target;
   GSList *osp_targets, *vts, *vt_groups;
@@ -2509,9 +2511,13 @@ launch_osp_openvas_task (task_t task, target_t target, const char *scan_id,
       else if (ret == -1)
         return -1;
       finished_hosts_str = report_finished_hosts_str (global_current_report);
+      clean_finished_hosts_str = clean_hosts_string (finished_hosts_str);
     }
   else
-    finished_hosts_str = NULL;
+    {
+      finished_hosts_str = NULL;
+      clean_finished_hosts_str = NULL;
+    }
 
   /* Set up target(s) */
   hosts_str = target_hosts (target);
@@ -2535,10 +2541,10 @@ launch_osp_openvas_task (task_t task, target_t target, const char *scan_id,
       gchar *new_exclude_hosts;
 
       new_exclude_hosts = g_strdup_printf ("%s,%s",
-                                           exclude_hosts_str,
-                                           finished_hosts_str);
-      free (exclude_hosts_str);
-      exclude_hosts_str = new_exclude_hosts;
+                                           clean_exclude_hosts,
+                                           clean_finished_hosts_str);
+      free (clean_exclude_hosts);
+      clean_exclude_hosts = new_exclude_hosts;
     }
 
   osp_target = osp_target_new (clean_hosts, ports_str, clean_exclude_hosts,
@@ -2553,6 +2559,7 @@ launch_osp_openvas_task (task_t task, target_t target, const char *scan_id,
   free (finished_hosts_str);
   g_free (clean_hosts);
   g_free (clean_exclude_hosts);
+  g_free (clean_finished_hosts_str);
   osp_targets = g_slist_append (NULL, osp_target);
 
   ssh_credential = target_osp_ssh_credential (target);
@@ -2843,6 +2850,7 @@ fork_osp_scan_handler (task_t task, target_t target, int from,
   switch (fork ())
     {
       case 0:
+        init_sentry ();
         break;
       case -1:
         /* Parent, failed to fork. */
@@ -2903,6 +2911,7 @@ fork_osp_scan_handler (task_t task, target_t target, int from,
 
       g_free (error);
       g_free (report_id);
+      gvm_close_sentry ();
       exit (-1);
     }
 
@@ -2934,6 +2943,7 @@ fork_osp_scan_handler (task_t task, target_t target, int from,
   set_scan_end_time_epoch (global_current_report, time (NULL));
   global_current_report = 0;
   current_scanner_task = (task_t) 0;
+  gvm_close_sentry ();
   exit (rc);
 }
 
@@ -3182,6 +3192,7 @@ fork_cve_scan_handler (task_t task, target_t target)
   switch (pid)
     {
       case 0:
+        init_sentry ();
         break;
       case -1:
         /* Parent, failed to fork. */
@@ -3223,6 +3234,7 @@ fork_cve_scan_handler (task_t task, target_t target)
                             "Error in target host list."
                             "  Interrupting scan.");
       set_report_scan_run_status (global_current_report, TASK_STATUS_INTERRUPTED);
+      gvm_close_sentry ();
       exit (1);
     }
 
@@ -3242,6 +3254,7 @@ fork_cve_scan_handler (task_t task, target_t target)
                               "  Interrupting scan.");
         set_report_scan_run_status (global_current_report, TASK_STATUS_INTERRUPTED);
         gvm_hosts_free (gvm_hosts);
+        gvm_close_sentry ();
         exit (1);
       }
   gvm_hosts_free (gvm_hosts);
@@ -3254,6 +3267,7 @@ fork_cve_scan_handler (task_t task, target_t target)
   set_report_scan_run_status (global_current_report, TASK_STATUS_DONE);
   global_current_report = 0;
   current_scanner_task = (task_t) 0;
+  gvm_close_sentry ();
   exit (0);
 }
 
@@ -4651,6 +4665,7 @@ scheduled_task_start (scheduled_task_t *scheduled_task,
         /* Restore the sigmask that was blanked for pselect. */
         pthread_sigmask (SIG_SETMASK, sigmask_current, NULL);
 
+        init_sentry ();
         reinit_manage_process ();
         manage_session_init (current_credentials.uuid);
         break;
@@ -4680,6 +4695,7 @@ scheduled_task_start (scheduled_task_t *scheduled_task,
         g_warning ("%s: fork_connection failed", __func__);
         reschedule_task (scheduled_task->task_uuid);
         scheduled_task_free (scheduled_task);
+        gvm_close_sentry ();
         exit (EXIT_FAILURE);
         break;
 
@@ -4708,6 +4724,7 @@ scheduled_task_start (scheduled_task_t *scheduled_task,
                              __func__,
                              scheduled_task->task_uuid);
                   scheduled_task_free (scheduled_task);
+                  gvm_close_sentry ();
                   exit (EXIT_FAILURE);
                 }
               if (errno == EINTR)
@@ -4720,6 +4737,7 @@ scheduled_task_start (scheduled_task_t *scheduled_task,
                          __func__,
                          scheduled_task->task_uuid);
               scheduled_task_free (scheduled_task);
+              gvm_close_sentry ();
               exit (EXIT_FAILURE);
             }
           if (WIFEXITED (status))
@@ -4771,6 +4789,7 @@ scheduled_task_start (scheduled_task_t *scheduled_task,
                       }
                   }
                   scheduled_task_free (scheduled_task);
+                  gvm_close_sentry ();
                   exit (EXIT_SUCCESS);
 
                 case EXIT_FAILURE:
@@ -4783,6 +4802,7 @@ scheduled_task_start (scheduled_task_t *scheduled_task,
           g_warning ("%s: child failed", __func__);
           reschedule_task (scheduled_task->task_uuid);
           scheduled_task_free (scheduled_task);
+          gvm_close_sentry ();
           exit (EXIT_FAILURE);
         }
     }
@@ -4801,6 +4821,7 @@ scheduled_task_start (scheduled_task_t *scheduled_task,
       g_warning ("%s: gmp_authenticate failed", __func__);
       scheduled_task_free (scheduled_task);
       gvm_connection_free (&connection);
+      gvm_close_sentry ();
       exit (EXIT_FAILURE);
     }
 
@@ -4822,6 +4843,7 @@ scheduled_task_start (scheduled_task_t *scheduled_task,
             g_warning ("%s: user denied permission to start task", __func__);
             scheduled_task_free (scheduled_task);
             gvm_connection_free (&connection);
+            gvm_close_sentry ();
             /* Return success, so that parent stops trying to start the task. */
             exit (EXIT_SUCCESS);
 
@@ -4829,12 +4851,14 @@ scheduled_task_start (scheduled_task_t *scheduled_task,
             g_warning ("%s: gmp_start_task and gmp_resume_task failed", __func__);
             scheduled_task_free (scheduled_task);
             gvm_connection_free (&connection);
+            gvm_close_sentry ();
             exit (EXIT_FAILURE);
         }
     }
 
   scheduled_task_free (scheduled_task);
   gvm_connection_free (&connection);
+  gvm_close_sentry ();
   exit (EXIT_SUCCESS);
 }
 
@@ -4891,6 +4915,7 @@ scheduled_task_stop (scheduled_task_t *scheduled_task,
     {
       scheduled_task_free (scheduled_task);
       gvm_connection_free (&connection);
+      gvm_close_sentry ();
       exit (EXIT_FAILURE);
     }
 
@@ -4898,11 +4923,13 @@ scheduled_task_stop (scheduled_task_t *scheduled_task,
     {
       scheduled_task_free (scheduled_task);
       gvm_connection_free (&connection);
+      gvm_close_sentry ();
       exit (EXIT_FAILURE);
     }
 
   scheduled_task_free (scheduled_task);
   gvm_connection_free (&connection);
+  gvm_close_sentry ();
   exit (EXIT_SUCCESS);
 }