+++ src/manage_sql_report_formats.c
@@ -23,6 +23,7 @@
  * The report format SQL for the GVM management layer.
  */
 
+#include "debug_utils.h"
 #include "manage_sql_report_formats.h"
 #include "manage_acl.h"
 #include "manage_report_formats.h"
@@ -43,6 +44,7 @@
 #include <sys/wait.h>
 #include <unistd.h>
 
+#include <gvm/base/gvm_sentry.h>
 #include <gvm/base/proctitle.h>
 #include <gvm/util/uuidutils.h>
 #include <gvm/util/fileutils.h>
@@ -3413,6 +3415,7 @@ run_report_format_script (gchar *report_format_id,
             {
               /* Child.  Drop privileges, run command, exit. */
 
+              init_sentry ();
               proctitle_set ("gvmd: Generating report");
 
               cleanup_manage_process (FALSE);
@@ -3421,6 +3424,7 @@ run_report_format_script (gchar *report_format_id,
                 {
                   g_warning ("%s (child): setgroups: %s",
                               __func__, strerror (errno));
+                  gvm_close_sentry ();
                   exit (EXIT_FAILURE);
                 }
               if (setgid (nobody->pw_gid))
@@ -3428,6 +3432,7 @@ run_report_format_script (gchar *report_format_id,
                   g_warning ("%s (child): setgid: %s",
                               __func__,
                               strerror (errno));
+                  gvm_close_sentry ();
                   exit (EXIT_FAILURE);
                 }
               if (setuid (nobody->pw_uid))
@@ -3435,6 +3440,7 @@ run_report_format_script (gchar *report_format_id,
                   g_warning ("%s (child): setuid: %s",
                               __func__,
                               strerror (errno));
+                  gvm_close_sentry ();
                   exit (EXIT_FAILURE);
                 }
 
@@ -3448,9 +3454,11 @@ run_report_format_script (gchar *report_format_id,
                               ret,
                               WEXITSTATUS (ret),
                               command);
+                  gvm_close_sentry ();
                   exit (EXIT_FAILURE);
                 }
 
+              gvm_close_sentry ();
               exit (EXIT_SUCCESS);
             }