+++ src/utils.c
@@ -49,6 +49,8 @@
 #include <sys/stat.h>
 #include <unistd.h>
 
+#include <gvm/base/gvm_sentry.h>
+
 #undef G_LOG_DOMAIN
 /**
  * @brief GLib log domain.
@@ -800,6 +802,7 @@ setup_signal_handler (int signal, void (*handler) (int), int block)
     {
       g_critical ("%s: failed to register %s handler",
                   __func__, strsignal (signal));
+      gvm_close_sentry ();
       exit (EXIT_FAILURE);
     }
 }
@@ -831,6 +834,7 @@ setup_signal_handler_info (int signal,
     {
       g_critical ("%s: failed to register %s handler",
                   __func__, strsignal (signal));
+      gvm_close_sentry ();
       exit (EXIT_FAILURE);
     }
 }